pyproject.toml
[tool.poetry]
name = "pokemon-generator"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
certifi = "^2021.5.30"
cycler = "^0.10.0"
future = "^0.18.2"
kiwisolver = "^1.3.1"
matplotlib = "^3.4.2"
numpy = "^1.20.3"
Pillow = "^8.2.0"
pyparsing = "^2.4.7"
python-dateutil = "^2.8.1"
six = "^1.16.0"
torch = "^1.8.1"
torchvision = "^0.9.1"
attrs = "^21.2.0"
jupyter = "^1.0.0"
jupyterlab = "^3.0.16"
pandas = "^1.2.4"
tqdm = "^4.61.0"
pytorch-fid = "^0.2.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
